Feature/wallet-refactor

+++ app/Actions/Wallet/UpdateTransactionAction.php
@@ -6,9 +6,13 @@ namespace App\Actions\Wallet;
 
 use App\Dtos\TransactionDto;
 use App\Enums\TransactionType;
+use App\Exceptions\WalletException;
+use App\Models\ChurchWallet;
+use Bavix\Wallet\Internal\Exceptions\RecordNotFoundException;
+use Bavix\Wallet\Internal\Exceptions\TransactionFailedException;
 use Bavix\Wallet\Models\Transaction;
-use Bavix\Wallet\Models\Wallet;
 use Bavix\Wallet\Services\FormatterService;
+use Illuminate\Support\Facades\DB;
 
 final readonly class UpdateTransactionAction
 {
@@ -18,34 +22,51 @@ final readonly class UpdateTransactionAction
         private FormatterService $formatterService,
     ) {}
 
-    public function handle(Transaction $transaction, TransactionDto $data, TransactionType $transactionType, ?Wallet $wallet = null): Transaction
+    public function handle(Transaction $transaction, TransactionDto $transactionDto, TransactionType $transactionType, ?ChurchWallet $wallet = null): Transaction
     {
-        $oldWallet = $transaction->wallet;
+        $oldWallet = ChurchWallet::find($transaction->wallet->holder_id);
         $isDeposit = $transactionType === TransactionType::DEPOSIT;
 
-        if (! $wallet instanceof Wallet) {
+        if (! $wallet instanceof ChurchWallet) {
             $wallet = $oldWallet;
         }
 
-        if ($oldWallet->id !== $wallet->id) {
-            $updatedTransaction = $isDeposit
-                ? $this->walletDepositAction->handle($wallet, $data)
-                : $this->walletWithdrawalAction->handle($wallet, $data);
-            $transaction->forceDelete();
-
-        } else {
-            $transaction->update([
-                'amount' => $this->formatterService->intValue($isDeposit ? $data->amount : -abs((float) $data->amount), 2),
-                'meta' => $data->meta->toArray(),
-                'confirmed' => $data->confirmed,
-            ]);
-            $updatedTransaction = $transaction->refresh();
-
+        if (! $oldWallet instanceof ChurchWallet || ! $wallet instanceof ChurchWallet) {
+            throw WalletException::notFound();
         }
 
-        $oldWallet->refreshBalance();
+        try {
+            return DB::transaction(function () use ($transaction, $transactionDto, $isDeposit, $wallet, $oldWallet): Transaction {
+                if ($oldWallet->id !== $wallet->id) {
+                    $transaction->forceDelete();
 
-        return $updatedTransaction;
+                    $updatedTransaction = $isDeposit
+                            ? $this->walletDepositAction->handle($wallet, $transactionDto)
+                            : $this->walletWithdrawalAction->handle($wallet, $transactionDto);
+                    $oldWallet->wallet->refreshBalance();
+
+                } else {
+                    $amount = $this->formatterService->intValue($isDeposit ? $transactionDto->amount : -abs((float) $transactionDto->amount), 2);
+                    if ($amount !== $transaction->amount) {
+                        $transaction->forceDelete();
+                        $oldWallet->wallet->refreshBalance();
+                        $updatedTransaction = $isDeposit
+                            ? $this->walletDepositAction->handle($wallet, $transactionDto)
+                            : $this->walletWithdrawalAction->handle($wallet, $transactionDto);
+                        $oldWallet->wallet->refreshBalance();
+                    } else {
+                        $updatedTransaction = $transaction;
+                    }
+
+                }
+
+                return $updatedTransaction;
+            });
+        } catch (RecordNotFoundException) {
+            throw WalletException::notFound();
+        } catch (TransactionFailedException) {
+            throw WalletException::transactionFailed();
+        }
 
     }
 }
